+++ src/job/files/mod.rs
@@ -44,8 +44,7 @@ async fn clean_empty_dirs() -> Result<(), anyhow::Error> {
         tokio::time::sleep(tokio::time::Duration::from_secs(3600)).await;
         let last_updated = std::time::SystemTime::now() - std::time::Duration::from_secs(3600);
         let root = format!("{}files/", config::get_config().common.data_wal_dir);
-        if let Err(e) =
-            config::utils::asynchronism::file::clean_empty_dirs(&root, Some(last_updated)).await
+        if let Err(e) = config::utils::async_file::clean_empty_dirs(&root, Some(last_updated)).await
         {
             log::error!("clean_empty_dirs, err: {}", e);
         }
